Build process: teach Grunt to show warnings but keep going
Without this patch, the Grunt process is too happy to stop due any issues
in the code. As many issues are caused by lack of intelligence in
TypeScript compiler, I think it's better to show warnings and keep going.

In long run, the code should be modified to not emit warnings but it
seems that even the official code is silently ignoring all warnings
about circular dependencies. This patch makes those visible but keeps
going unless unknown warnings or real errors have been encountered.

This patch also makes all linter errors about CSS/Less non-critical
because it seems that the issues with the current source are more about
stupid tools instead of real issues. Here, too, the correct long term
solution is to improve tools or change to simpler syntax that the
existing tools can handle correctly.

Of course, applying this patch means that you have to actually read the
logs instead of thinking that finished build means everything is good.